% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsReg.R
\name{lsReg}
\alias{lsReg}
\title{Title}
\usage{
lsReg(basemdl, colstoadd, testtype)
}
\arguments{
\item{basemdl}{Base model, y ~ xl, fitted with glm}

\item{colstoadd}{Number of columns in xr. Full model will be y ~ xl + xr}
}
\value{
The data set
}
\description{
Title
}
\examples{
linmdlfile <- system.file("extdata", "linmodel.rds", package = "lsReg")
linmodel <- readRDS(linmdlfile)
link <- lsReg(linmodel, 1)
}
