% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststat.R
\name{runtest}
\alias{runtest}
\title{Title}
\usage{
runtest(lsregmem, xr)
}
\arguments{
\item{lsregmem}{Memory allocated for large scale regression}

\item{xr}{Additional variables to be tested}
}
\value{
Nonzero value indicates an error
}
\description{
Title
}
